[hb-ft] Adjust serial signature

+++ src/hb-ft.cc
@@ -86,7 +86,7 @@ struct hb_ft_font_t
 
   mutable hb_mutex_t lock;
   FT_Face ft_face;
-  mutable unsigned long cached_serial;
+  mutable unsigned cached_serial;
   mutable hb_advance_cache_t advance_cache;
 };
 
@@ -103,7 +103,7 @@ _hb_ft_font_create (FT_Face ft_face, bool symbol, bool unref)
 
   ft_font->load_flags = FT_LOAD_DEFAULT | FT_LOAD_NO_HINTING;
 
-  ft_font->cached_serial = (unsigned long) -1;
+  ft_font->cached_serial = (unsigned) -1;
   ft_font->advance_cache.init ();
 
   return ft_font;
